requirements.txt
flask==1.0.2
flask-sqlalchemy==2.3.0
psycopg2==2.8.4
datetime
uuid
PyJWT==1.7.1
Werkzeug